move and update model manager
